requirements.txt
numpy==1.25.2
pandas==2.3.1
PyYAML==6.0.2
